QL Quỹ

+++ Models/Quy.cs
@@ -11,8 +11,8 @@ namespace QuanLyBanHang.Models
         public int Id { get; set; }
         public string Name { get; set; }
         public string Type { get; set; }
-        public double Balance { get; set; }
-        public double TotalTransaction { get; set; }
+        public int Balance { get; set; }
+        public int TotalTransaction { get; set; }
         public string Status { get; set; }
     }
 }
